backend/requirements.txt
Django==4.2.14
django-cors-headers==4.4.0
djangorestframework==3.15.2
python-dotenv==1.0.1
requests==2.31.0